requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
plotly
langchain
langchain-groq
python-dotenv
langchain-experimental
tabulate
joblib
fpdf